refactor: modify the structure of the repository to be out of domain package
